Modify player with new message protocol

+++ avoidingproblems/src/main/java/sdc/avoidingproblems/circuits/algebra/Util.java
@@ -12,7 +12,7 @@ public class Util {
 
    public static FieldElement getFieldElementInstance(Class<?> clazz, Long value, Long MOD) throws ClassNotSupportedException {
       try {
-         Constructor<?> constructor = clazz.getConstructor(Integer.class, Integer.class);
+         Constructor<?> constructor = clazz.getConstructor(Long.class, Long.class);
          FieldElement result = (FieldElement) constructor.newInstance(value, MOD);
          return result;
       } catch (NoSuchMethodException | SecurityException | InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException ex) {
